requirements.txt
Flask==2.1.2
Flask_MySQLdb==1.0.1
keras==2.9.0
numpy==1.22.4
opencv_python==4.5.5.64
Pillow==9.1.1
PyYAML==6.0
